// task8/dapp-task8/src/types/SimpleCalculator.d.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from 'bn.js';
import { ContractOptions } from 'web3-eth-contract';
import { EventLog } from 'web3-core';
import { EventEmitter } from 'events';
import {
    Callback,
    PayableTransactionObject,
    NonPayableTransactionObject,
    BlockType,
    ContractEventLog,
    BaseContract
} from './types';

interface EventOptions {
    filter?: object;
    fromBlock?: BlockType;
    topics?: string[];
}

export interface SimpleCalculator extends BaseContract {
    constructor(
        jsonInterface: any[],
        address?: string,
        options?: ContractOptions
    ): SimpleCalculator;
    clone(): SimpleCalculator;
    methods: {
        add(x: number | string | BN): PayableTransactionObject<void>;

        sub(x: number | string | BN): PayableTransactionObject<void>;

        getResult(): NonPayableTransactionObject<string>;
    };
    events: {
        allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
    };
}